handle command errors
