+++ src/main/java/tasks/Deadline.java
@@ -1,3 +1,5 @@
+package tasks;
+
 public class Deadline extends Task {
 
     private String by;
