+++ src/main/java/tasks/Event.java
@@ -1,4 +1,6 @@
-public class Event extends Task{
+package tasks;
+
+public class Event extends Task {
 
     private String at;
